Fix crash when using CALL command with syscalls.
The unused syscalls were commented out, but the CALL command relied on
all syscalls being enumerated sequentially. Uncomment the syscalls and
leave unknown calls with generic names.

Also, fix the range for which CALL translates the address to a syscall,
since it did not include the full range at the upper end.