Fix #7: Use m68k-elf-gcc in GitHub action
The Linux ABI for m68k allows unaligned memory accesses (i.e. a 68020 or
better is required). Since the AlphaSmart OS doesn't enable this (due to
backward compatibility with the 68000-based AS3000), we need to use a
non-Linux toolchain so that gcc respects -mstrict-align (which is the
default).

This also required providing prototypes for the C library functions
provided by the OS, which was on the todo list anyway.