Fix EhBASIC linking order and reduce RAM size
The must be linked in a certain order, or else the assembly code tries
to overwrite the global data of the C code.

Also, the 32kB RAM was excessive as this is mostly for demonstration
purposes. If more RAM is required, a method to use free memory could
be investigated.