Simulate 32-bit offset calculation
Reserve register D7 for the calculation of 32-bit offsets from the
program counter; this permits programs larger than 64KiB.

Currently, no link-time relaxation is implemented. The size appears to
increase by roughly 10%, and of course there is also a slight time
penalty for the extended calculations.

As a side effect, an applet's RAM area could also now be larger than
32KiB, via the same mechanism.